server/requirements.txt
Flask 
google-cloud-firestore
python-dotenv
flask-cors
openai